add more options to test push policy command

+++ blaze/command/test.py
@@ -18,6 +18,12 @@ from . import command
 
 
 @command.argument("url", nargs="?", help="The URL to analyze the page load time for")
+@command.argument("--from_manifest", help="The training manifest file to use as input to the simulator")
+@command.argument(
+    "--only_simulator",
+    action="store_true",
+    help="Only evaluate the page load time on the simulator (must be loaded from manifest to use this)",
+)
 @command.argument("--policy_type", help="The test type to run", choices=["simple", "random"])
 @command.argument(
     "--random_chance",
@@ -33,12 +39,21 @@ def test_push(args):
     """
     Runs a pre-defined test on the given webpage
     """
-    if args.policy_type == "simple":
-        _test_push(args.url, 1, _simple_push_policy_generator(), args.bandwidth, args.latency)
-    if args.policy_type == "random":
-        _test_push(
-            args.url, args.iterations, _random_push_policy_generator(args.random_chance), args.bandwidth, args.latency
-        )
+    simple_policy = args.policy_type == "simple"
+    policy_generator = (
+        _simple_push_policy_generator() if simple_policy else _random_push_policy_generator(args.random_chance)
+    )
+    _test_push(
+        **{
+            "url": args.url,
+            "manifest": args.manifest,
+            "iterations": 1 if simple_policy else args.iterations,
+            "policy_generator": policy_generator,
+            "bandwidth": args.bandwidth,
+            "latency": args.latency,
+            "only_simulator": args.only_simulator,
+        }
+    )
 
 
 def _simple_push_policy_generator() -> Callable[[List[PushGroup]], Policy]:
@@ -79,57 +94,50 @@ def _random_push_policy_generator(chance: float) -> Callable[[List[PushGroup]], 
 
 
 def _test_push(
-    url: str,
+    url: Optional[str],
+    manifest: Optional[str],
     iterations: int,
     policy_generator: Callable[[List[PushGroup]], Policy],
     bandwidth: Optional[int],
     latency: Optional[int],
+    only_simulator: Optional[bool],
 ):
     default_client_env = get_default_client_environment()
     client_env = get_client_environment_from_parameters(
         bandwidth or default_client_env.bandwidth, latency or default_client_env.latency, 1
     )
 
-    log.info("calculating page load time", url=url)
-    with tempfile.TemporaryDirectory() as record_dir:
-        # this is to work around the fact that mahimahi needs an empty directory
-        # so we use TemporaryDirectory to get a unique name for a directory and
-        # then delete it. After mahimahi runs and create the dir, then TemporaryDirectory
-        # can delete it as normal
-        os.rmdir(record_dir)
+    if not only_simulator:
+        log.info("calculating page load time", url=url)
+        with tempfile.TemporaryDirectory() as record_dir:
+            if not manifest:
+                # this is to work around the fact that mahimahi needs an empty directory
+                # so we use TemporaryDirectory to get a unique name for a directory and
+                # then delete it. After mahimahi runs and create the dir, then TemporaryDirectory
+                # can delete it as normal
+                os.rmdir(record_dir)
 
-        config = get_config(EnvironmentConfig(replay_dir=record_dir, request_url=url))
-        log.debug("recording webpage in Mahimahi", record_dir=record_dir)
-        record_webpage(url, record_dir, config)
+                config = get_config(EnvironmentConfig(replay_dir=record_dir, request_url=url))
+                log.debug("recording webpage in Mahimahi", record_dir=record_dir)
+                record_webpage(url, record_dir, config)
+            else:
+                config = get_config(EnvironmentConfig.load_file(manifest))
+                log.debug("using recorded webpage", record_dir=config.env_config.replay_dir)
 
-        log.debug("capturing median PLT in mahimahi with given environment")
-        plt, res_list, push_groups = get_page_load_time_in_mahimahi(config.env_config.request_url, client_env, config)
+            log.debug("capturing median PLT in mahimahi with given environment")
+            plt, res_list, push_groups, push_plts, push_policies = _get_results_in_mahimahi(
+                config, client_env, iterations, policy_generator, capture_default=bool(bandwidth or latency)
+            )
 
-        # If the user passed in a custom environment, we want to use the PLT from that environment
-        # but we want to use the HAR from the default page load to run in the simulator. This is to
-        # allow the simulator to simulate the custom environment on top of the default environment
-        # and to prevent conflating environments
-        if bandwidth or latency:
-            log.debug("capturing HAR in mahimahi for simulator in default environment")
-            default_har = capture_har_in_mahimahi(url, config, default_client_env)
-            res_list = har_entries_to_resources(default_har)
-            push_groups = resource_list_to_push_groups(res_list)
+            env_config = EnvironmentConfig(
+                replay_dir=config.env_config.replay_dir,
+                request_url=config.env_config.request_url,
+                push_groups=push_groups,
+                har_resources=res_list,
+            )
 
-        push_plts = []
-        push_policies = []
-
-        for _ in range(iterations):
-            policy = policy_generator(push_groups)
-            push_policies.append(policy)
-
-            log.debug("getting HAR in mahimahi with push policy:")
-            log.debug(json.dumps(policy.as_dict, indent=4))
-            push_plt, *_ = get_page_load_time_in_mahimahi(config.env_config.request_url, client_env, config, policy)
-            push_plts.append(push_plt)
-
-        env_config = EnvironmentConfig(
-            replay_dir=record_dir, request_url=url, push_groups=push_groups, har_resources=res_list
-        )
+    else:
+        env_config = EnvironmentConfig.load_file(manifest)
 
     log.debug("running simulator...")
     sim = Simulator(env_config)
@@ -140,3 +148,33 @@ def _test_push(
     log.info("real push page load times", page_load_time=push_plts)
     log.info("simulated page load time", page_load_time=round(sim_plt, 3))
     log.info("simulated push page load time", page_load_time=[round(plt, 3) for plt in push_sim_plts])
+
+
+def _get_results_in_mahimahi(config, client_env, iterations, policy_generator, capture_default=False):
+    log.debug("capturing median PLT in mahimahi with given environment")
+    plt, res_list, push_groups = get_page_load_time_in_mahimahi(config.env_config.request_url, client_env, config)
+
+    # If the user passed in a custom environment, we want to use the PLT from that environment
+    # but we want to use the HAR from the default page load to run in the simulator. This is to
+    # allow the simulator to simulate the custom environment on top of the default environment
+    # and to prevent conflating environments
+    if capture_default:
+        default_client_env = get_default_client_environment()
+        log.debug("capturing HAR in mahimahi for simulator in default environment")
+        default_har = capture_har_in_mahimahi(config.env_config.request_url, config, default_client_env)
+        res_list = har_entries_to_resources(default_har)
+        push_groups = resource_list_to_push_groups(res_list)
+
+    push_plts = []
+    push_policies = []
+
+    for _ in range(iterations):
+        policy = policy_generator(push_groups)
+        push_policies.append(policy)
+
+        log.debug("getting HAR in mahimahi with push policy:")
+        log.debug(json.dumps(policy.as_dict, indent=4))
+        push_plt, *_ = get_page_load_time_in_mahimahi(config.env_config.request_url, client_env, config, policy)
+        push_plts.append(push_plt)
+
+    return plt, res_list, push_groups, push_plts, push_policies
